test(libeval): split orchestration-toolkit/facilitator/redaction/trace-query by behaviour family
- orchestration-toolkit.test.js (498) → -handlers (simple handlers +
  server factories) and -asks (Ask/Answer/cancel/remind), shared stubBus
  lifted to -helpers.js.
- facilitator.test.js (493) → -orchestration (core orchestration) and
  -messaging (messaging + bidirectional Ask), shared msg/seed helpers
  lifted to -helpers.js.
- redaction.test.js (481) → -matching (env allowlist, patterns, benign,
  word-boundary) and -opt-out (opt-out, noop, exports), shared
  sentinel/stderr helpers lifted to -helpers.js.
- trace-query.test.js (460) → -navigation (overview/count/batch/head/
  tail/search/createTraceQuery) and -analysis (toolFrequency/tool/
  errors/reasoning/timeline/stats), shared buildTrace lifted to -helpers.js.

Assertions unchanged; test counts preserved (26, 8, 22, 30).
Spec 0640 § C (large files), design Decision 5.
