+++ libraries/libeval/test/redaction-matching.test.js
@@ -1,47 +1,11 @@
 import { describe, test } from "node:test";
 import assert from "node:assert";
-import { createDefaultRuntime } from "@forwardimpact/libutil/runtime";
-const _rt = createDefaultRuntime();
 
+import { createRedactor } from "../src/redaction.js";
 import {
-  Redactor,
-  createRedactor,
-  createNoopRedactor,
-  DEFAULT_ENV_ALLOWLIST,
-  DEFAULT_PATTERNS,
-} from "../src/redaction.js";
-
-/**
- * Guard helper: sentinels must be JSON-stable (printable ASCII without `"`,
- * `\`, or control characters) so a substring scan over JSON-encoded bytes
- * gives a sound check. A non-stable sentinel would JSON-escape and pass
- * the substring check even when redaction missed.
- */
-function assertJsonStableSentinel(s) {
-  // biome-ignore lint/suspicious/noControlCharactersInRegex: validating absence of control chars in test sentinels
-  if (/[\x00-\x1f\x7f"\\]/.test(s)) {
-    throw new Error(`sentinel is not JSON-stable: ${JSON.stringify(s)}`);
-  }
-}
-
-/**
- * Capture stderr writes from a callback. Restores the original write
- * regardless of throw outcome.
- */
-function captureStderr(fn) {
-  const captured = [];
-  const orig = process.stderr.write.bind(process.stderr);
-  process.stderr.write = (chunk) => {
-    captured.push(String(chunk));
-    return true;
-  };
-  try {
-    fn();
-  } finally {
-    process.stderr.write = orig;
-  }
-  return captured.join("");
-}
+  rt as _rt,
+  assertJsonStableSentinel,
+} from "./redaction-helpers.js";
 
 describe("Redactor — env-var allowlist (criterion 1)", () => {
   test("replaces sentinels with [REDACTED:env:NAME] across deep-walked carrier shapes", () => {
@@ -278,125 +242,6 @@ describe("Redactor — credential patterns (criterion 2)", () => {
   });
 });
 
-describe("Redactor — benign content unchanged (criterion 3)", () => {
-  const r = createRedactor({ runtime: _rt, env: {} });
-  const benign = [
-    "Hello world — this is plain prose.",
-    "# Markdown header\n\n- item 1\n- item 2",
-    "https://www.forwardimpact.team/docs/products/index.md",
-    "Visit https://github.com/forwardimpact/monorepo/pull/123.",
-    // git SHA (40 hex)
-    "7dd76efba1234567890abcdef0123456789abcde",
-    // UUID
-    "550e8400-e29b-41d4-a716-446655440000",
-    // ghp_ prefix at less than 36 chars — should NOT match
-    "ghp_short",
-    "ghp_" + "A".repeat(35),
-    // quoted shell commands
-    "echo 'hello world' | grep -v foo",
-    'curl -X POST -d "{\\"foo\\":1}" http://example.com',
-  ];
-  for (const text of benign) {
-    test(`round-trips identically: ${JSON.stringify(text).slice(0, 60)}`, () => {
-      assert.strictEqual(r.redactValue(text), text);
-    });
-  }
-});
-
-describe("Redactor — opt-out (criterion 4, design § Opt-out surface)", () => {
-  test("LIBEVAL_REDACTION_DISABLED=1 disables and emits stderr warning exactly once", () => {
-    let r;
-    const stderr = captureStderr(() => {
-      r = createRedactor({
-        runtime: _rt,
-        env: {
-          LIBEVAL_REDACTION_DISABLED: "1",
-          GH_TOKEN: "would-have-redacted",
-        },
-      });
-    });
-    assert.strictEqual(r.enabled, false);
-    assert.strictEqual(
-      r.redactValue("would-have-redacted"),
-      "would-have-redacted",
-    );
-    assert.match(stderr, /libeval: trace redaction DISABLED/);
-    // Single warning per construction.
-    const matches = stderr.match(/redaction DISABLED/g) ?? [];
-    assert.strictEqual(matches.length, 1);
-  });
-
-  test('LIBEVAL_REDACTION_DISABLED="true" does NOT disable (literal "1" is the contract)', () => {
-    let r;
-    const stderr = captureStderr(() => {
-      r = createRedactor({
-        runtime: _rt,
-        env: { LIBEVAL_REDACTION_DISABLED: "true", GH_TOKEN: "secret-value" },
-      });
-    });
-    assert.strictEqual(r.enabled, true);
-    assert.strictEqual(
-      r.redactValue("secret-value"),
-      "[REDACTED:env:GH_TOKEN]",
-    );
-    assert.strictEqual(stderr, "");
-  });
-
-  test('LIBEVAL_REDACTION_DISABLED="yes" does NOT disable (literal "1" is the contract)', () => {
-    let r;
-    const stderr = captureStderr(() => {
-      r = createRedactor({
-        runtime: _rt,
-        env: { LIBEVAL_REDACTION_DISABLED: "yes", GH_TOKEN: "secret-value" },
-      });
-    });
-    assert.strictEqual(r.enabled, true);
-    assert.strictEqual(
-      r.redactValue("secret-value"),
-      "[REDACTED:env:GH_TOKEN]",
-    );
-    assert.strictEqual(stderr, "");
-  });
-
-  test("createRedactor({ enabled: false }) fires the stderr warning regardless of env state", () => {
-    let r;
-    const stderr = captureStderr(() => {
-      r = createRedactor({ runtime: _rt, env: {}, enabled: false });
-    });
-    assert.strictEqual(r.enabled, false);
-    assert.match(stderr, /libeval: trace redaction DISABLED/);
-  });
-
-  test("disabled redactor returns top-level input by reference (identity contract)", () => {
-    const r = createNoopRedactor();
-    const obj = { type: "assistant", message: { content: [{ text: "hi" }] } };
-    assert.strictEqual(r.redactValue(obj), obj);
-    const arr = [1, "two", { three: 3 }];
-    assert.strictEqual(r.redactValue(arr), arr);
-    assert.strictEqual(r.redactValue("plain"), "plain");
-  });
-});
-
-describe("createNoopRedactor", () => {
-  test("returns a Redactor whose redactValue is identity", () => {
-    const r = createNoopRedactor();
-    assert.ok(r instanceof Redactor);
-    assert.strictEqual(r.enabled, false);
-    const v = { a: "x" };
-    assert.strictEqual(r.redactValue(v), v);
-  });
-
-  test("never fires the stderr warning regardless of env state", () => {
-    // Even though disabled, the noop helper must NOT write to stderr —
-    // it is intended for test fixtures that need a silent disabled
-    // redactor.
-    const stderr = captureStderr(() => {
-      createNoopRedactor();
-    });
-    assert.strictEqual(stderr, "");
-  });
-});
-
 describe("Redactor — word boundary adversarial cases (Risks table)", () => {
   const r = createRedactor({ runtime: _rt, env: {} });
   const body = "A".repeat(36);
@@ -435,47 +280,5 @@ describe("Redactor — word boundary adversarial cases (Risks table)", () => {
   test("ghp_<37> (one extra word char) does NOT match (anchored to 36)", () => {
     const longer = `ghp_${"A".repeat(37)}`;
     assert.strictEqual(r.redactValue(longer), longer);
-  });
-});
-
-describe("Redactor — exports and defaults", () => {
-  test("DEFAULT_ENV_ALLOWLIST is the documented contract", () => {
-    assert.deepStrictEqual(
-      [...DEFAULT_ENV_ALLOWLIST],
-      [
-        "ANTHROPIC_API_KEY",
-        "AWS_ACCESS_KEY_ID",
-        "AWS_SECRET_ACCESS_KEY",
-        "DATABASE_PASSWORD",
-        "GH_TOKEN",
-        "GITHUB_TOKEN",
-        "JWT_SECRET",
-        "MCP_TOKEN",
-        "MICROSOFT_APP_ID",
-        "MICROSOFT_APP_PASSWORD",
-        "MICROSOFT_APP_TENANT_ID",
-        "PRODUCT_LANDMARK_TOKEN",
-        "SERVICE_SECRET",
-        "SUPABASE_ANON_KEY",
-        "SUPABASE_SERVICE_ROLE_KEY",
-      ],
-    );
-  });
-
-  test("DEFAULT_PATTERNS covers the five documented kinds", () => {
-    const kinds = DEFAULT_PATTERNS.map((p) => p.kind);
-    assert.deepStrictEqual(kinds, [
-      "anthropic",
-      "gh-pat",
-      "gh-installation",
-      "gh-oauth",
-      "gh-fine-grained",
-    ]);
-  });
-
-  test("createRedactor({ runtime: _rt }) with no options falls back to process.env", () => {
-    // Smoke check — must not throw, and must produce a Redactor.
-    const r = createRedactor({ runtime: _rt });
-    assert.ok(r instanceof Redactor);
   });
 });
